color project admin panel done

+++ app/Http/Controllers/Admin/ProductController.php
@@ -11,10 +11,7 @@ use Illuminate\Support\Facades\Storage;
 
 class ProductController extends Controller
 {
-    //
-
     public function index(?int $id = null){
-
        
         $searchValue = request()->query("search",null);
 
@@ -31,27 +28,28 @@ class ProductController extends Controller
             $products = Product::with('category')->simplePaginate(3);
         }; 
 
-     
-        
-        
         return view('admin.productView',compact(['categories','products','editItem']));
     }
 
 
 
-
     public function store(Request $request,?int $id=null){
 
-
-        $request->validate([
+        $validationRules = [
             'name'=> 'required',
             'category_id'=> 'required',
+            'price'=> 'required|int',
             'description'=> 'required',
-            'picture' => 'nullable|image|mimes:jpeg,jpg,png,gif,webp,svg'
+        ] ;
 
-        ]);
+        if($id == null){
+            $validationRules['picture'] = 'required|image|mimes:jpeg,jpg,png,gif,webp,svg';
+        }
 
-        $data = $request->only(['name','description','logn_description','category_id']);
+        $request->validate($validationRules);
+        $data = $request->only(['name','description','price','logn_description','category_id']);
+
+        // return response()->json($data);
 
         if(!is_null($id)){
 
@@ -70,57 +68,28 @@ class ProductController extends Controller
             return redirect()->route('admin.product',['page'=>request()->query('page'),'search'=>request()->query('search')])->with('success','Successfully edit');
         }
 
-
         if($request->hasFile('picture')){
              $path = $request->file('picture')->store('productthum');
             $data['picture'] = $path;
         }
-
         //creating product
         $product = Product::create($data);
-
-        //creating image from product
-        
-        // if($request->hasFile('img')){
-        //     $allFile = $request->file('img');
-        //     foreach($allFile as $file){
-        //         $path = $file->store('product');
-        //         ProductImage::create([
-        //             'product_id' => $product->id,
-        //             'img' => $path
-        //         ]);
-        //     }
-        // }
-
-        // return response()->json([
-        //     "success" => true,
-        // ]);
 
         return redirect()->back()->with('success','Successfully created Product');
     }
 
-
     public function destory(int $id){
-        
-        $allimage = ProductImage::where('product_id','=',$id)->get();
-        foreach($allimage as $image){
-            //unlink image...
-            Storage::delete($image->img);
-            $image->delete();
-        }
-
-        
         $deleteProduct = Product::find($id);
         if($deleteProduct){
-            Storage::delete($deleteProduct->picture);
+            if($deleteProduct->picture){
+                Storage::delete($deleteProduct->picture);
+            }
+            
             $deleteProduct->delete();
         }
         return redirect()->route('admin.product')->with('success','successfully deleted');
-        
     }
 
-
-
-
+    
 
 }
